pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hand-to-excalidraw"
version = "0.1.0"
description = "Convert hand-drawn flowcharts to Excalidraw files using Qwen2.5-VL"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "huggingface-hub>=0.25.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "Pillow>=10.0.0",
]

[project.scripts]
hand2excalidraw = "app.cli:main"

[tool.setuptools.packages.find]
include = ["app*"]
